Refactor subproject to workstation.gui
