+++ gui/icons/icon.png
